+++ corda-testacles-cordapp-example/src/test/kotlin/com/github/manosbatsis/corda/testacles/nodedriver/NodeDriverHelperClassLfTest.kt
@@ -30,6 +30,7 @@ import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeAll
+import org.junit.jupiter.api.Tag
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS
@@ -37,6 +38,7 @@ import org.slf4j.LoggerFactory
 
 /** Sample class lifecycle test using the [NodeDriverHelper] directly */
 @TestInstance(PER_CLASS)
+@Tag("nodedriver")
 class NodeDriverHelperClassLfTest {
 
     companion object {